qa/requirements.txt
docker-py
jinja2
pyaml
pyOpenSSL
ndg-httpsclient
pyasn1
